Clean up network state logic
